Support deriving CLike
Fixes #22
